requirements.txt
django~=1.9.0
django-xadmin
django-crispy-forms~=1.6.0
django-reversion~=2.0.0
django-formtools==1.0
future==0.15.2
httplib2==0.9.2
six==1.10.0
Pillow==2.0.0
requests==1.2.0
yolk==0.4.3
